pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "leropilot"
version = "0.1.0"
description = "LeRoPilot - Local development environment manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "structlog==24.4.0",
    "pyyaml==6.0.2",
    "httpx==0.28.1",
    "pywinpty>=2.0.0; sys_platform == 'win32'",
    "pyserial>=3.5",
    "opencv-python>=4.12.0.88",
    "dynamixel-sdk>=4.0.2",
    "feetech-servo-sdk>=1.0.0",
    "python-can>=4.0.0",
    "pyrealsense2>=2.50.0",
    "uptime>=3.0.1",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.16.0",
    "mypy>=1.8.0,<2.0.0",
    "ruff>=0.6.0,<1.0.0",
    "types-PyYAML>=6.0.0,<7.0.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
leropilot = ["static/**/*"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# NOTE: flake8 and isort configuration removed — ruff is the primary linter.
# If you still need flake8/isort, re-add configuration sections or install
# corresponding tools. Ruff is configured in [tool.ruff] above and includes
# import-sorting rules (select = ["I"]).

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "build",
    "dist",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "ANN",    # flake8-annotations
]
extend-ignore = [
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
